[backend/Frontend] Display and create Triggers in Settings > User
