+++ opencti-platform/opencti-front/src/private/components/settings/users/Root.jsx
@@ -1,14 +1,10 @@
-import React, { Component } from 'react';
+import React, { useEffect, useMemo } from 'react';
 import * as PropTypes from 'prop-types';
-import { Route, withRouter } from 'react-router-dom';
-import { graphql } from 'react-relay';
-import {
-  QueryRenderer,
-  requestSubscription,
-} from '../../../../relay/environment';
+import { Route, Switch, useParams } from 'react-router-dom';
+import { graphql, usePreloadedQuery, useQueryLoader, useSubscription } from 'react-relay';
 import TopBar from '../../nav/TopBar';
 import User, { userQuery } from './User';
-import Loader from '../../../../components/Loader';
+import Loader, { LoaderVariant } from '../../../../components/Loader';
 import ErrorNotFound from '../../../../components/ErrorNotFound';
 import { SETTINGS_SETACCESSES } from '../../../../utils/hooks/useGranted';
 import Security from '../../../../utils/Security';
@@ -46,72 +42,73 @@ const subscription = graphql`
   }
 `;
 
-class RootUser extends Component {
-  constructor(props) {
-    super(props);
-    const {
-      match: {
-        params: { userId },
-      },
-    } = props;
-    this.sub = requestSubscription({
+const RootUserComponent = ({ queryRef, userId, refetch }) => {
+  const subConfig = useMemo(
+    () => ({
       subscription,
       variables: { id: userId },
-    });
-  }
+    }),
+    [userId],
+  );
+  useSubscription(subConfig);
+  const data = usePreloadedQuery(userQuery, queryRef);
+  const { user } = data;
 
-  componentWillUnmount() {
-    this.sub.dispose();
-  }
+  return (
+    <Security needs={[SETTINGS_SETACCESSES]}>
+      {user ? (
+        <Switch>
+          <Route
+            exact
+            path="/dashboard/settings/accesses/users/:userId"
+            render={(routeProps) => (
+              <User {...routeProps} userData={user} refetch={refetch}/>
+            )}
+          />
+        </Switch>
+      ) : (
+        <ErrorNotFound />
+      )}
+    </Security>
+  );
+};
 
-  render() {
-    const {
-      match: {
-        params: { userId },
-      },
-    } = this.props;
-    return (
-      <Security needs={[SETTINGS_SETACCESSES]}>
-        <TopBar />
-        <QueryRenderer
-          query={userQuery}
-          variables={{
-            id: userId,
-            rolesOrderBy: 'name',
-            rolesOrderMode: 'asc',
-            groupsOrderBy: 'name',
-            groupsOrderMode: 'asc',
-            organizationsOrderBy: 'name',
-            organizationsOrderMode: 'asc',
-          }}
-          render={({ props }) => {
-            if (props) {
-              if (props.user) {
-                return (
-                  <div>
-                    <Route
-                      exact
-                      path="/dashboard/settings/accesses/users/:userId"
-                      render={(routeProps) => (
-                        <User {...routeProps} user={props.user} />
-                      )}
-                    />
-                  </div>
-                );
-              }
-              return <ErrorNotFound />;
-            }
-            return <Loader />;
-          }}
-        />
-      </Security>
-    );
-  }
-}
+const RootUser = () => {
+  const { userId } = useParams();
+  const queryParams = {
+    id: userId,
+    rolesOrderBy: 'name',
+    rolesOrderMode: 'asc',
+    groupsOrderBy: 'name',
+    groupsOrderMode: 'asc',
+    organizationsOrderBy: 'name',
+    organizationsOrderMode: 'asc' };
+  const [queryRef, loadQuery] = useQueryLoader(userQuery);
+  useEffect(() => {
+    loadQuery(queryParams, { fetchPolicy: 'store-and-network' });
+  }, []);
+
+  const refetch = React.useCallback(() => {
+    loadQuery(queryParams, { fetchPolicy: 'store-and-network' });
+  }, [queryRef]);
+
+  return (
+    <div>
+      <TopBar />
+      {queryRef ? (
+        <React.Suspense fallback={<Loader variant={LoaderVariant.inElement} />}>
+          <RootUserComponent queryRef={queryRef} userId={userId} refetch={refetch}/>
+        </React.Suspense>
+      ) : (
+        <Loader variant={LoaderVariant.inElement} />
+      )}
+    </div>
+  );
+};
 
 RootUser.propTypes = {
   children: PropTypes.node,
   match: PropTypes.object,
 };
 
-export default withRouter(RootUser);
+export default RootUser;
