+++ opencti-platform/opencti-graphql/src/database/engine.js
@@ -71,7 +71,13 @@ import { isStixObject } from '../schema/stixCoreObject';
 import { isBasicRelationship, isStixRelationshipExceptRef } from '../schema/stixRelationship';
 import { RELATION_INDICATES } from '../schema/stixCoreRelationship';
 import { INTERNAL_FROM_FIELD, INTERNAL_TO_FIELD } from '../schema/identifier';
-import { computeUserMemberAccessIds, isBypassUser, MEMBER_ACCESS_ALL } from '../utils/access';
+import {
+  computeUserMemberAccessIds, INTERNAL_USERS,
+  isBypassUser,
+  isUserHasCapability,
+  MEMBER_ACCESS_ALL,
+  SETTINGS_SET_ACCESSES
+} from '../utils/access';
 import { isSingleRelationsRef, } from '../schema/stixEmbeddedRelationship';
 import { now, runtimeFieldObservableValueScript } from '../utils/format';
 import { ENTITY_TYPE_MARKING_DEFINITION } from '../schema/stixMetaObject';
@@ -273,7 +279,9 @@ const buildDataRestrictions = async (context, user, opts = {}) => {
   // eslint-disable-next-line camelcase
   const must_not = [];
   // check user access
-  must.push(...buildUserMemberAccessFilter(user, opts?.adminBypassUserAccess));
+  if (!INTERNAL_USERS[user.id]) {
+    must.push(...buildUserMemberAccessFilter(user, opts?.adminBypassUserAccess));
+  }
   // If user have bypass, no need to check restrictions
   if (!isBypassUser(user)) {
     // region Handle marking restrictions
@@ -377,8 +385,8 @@ const buildDataRestrictions = async (context, user, opts = {}) => {
   return { must, must_not };
 };
 
-export const buildUserMemberAccessFilter = (user, adminBypassUserAccess = true) => {
-  if (adminBypassUserAccess && isBypassUser(user)) {
+export const buildUserMemberAccessFilter = (user, adminBypassUserAccess = false) => {
+  if (adminBypassUserAccess && isUserHasCapability(user, SETTINGS_SET_ACCESSES)) {
     return [];
   }
   const userAccessIds = computeUserMemberAccessIds(user);
@@ -881,7 +889,7 @@ export const elFindByIds = async (context, user, ids, opts = {}) => {
       };
       mustTerms.push(shouldType);
     }
-    const markingRestrictions = await buildDataRestrictions(context, user);
+    const markingRestrictions = await buildDataRestrictions(context, user, { adminBypassUserAccess: true });
     mustTerms.push(...markingRestrictions.must);
     const body = {
       query: {
@@ -1060,7 +1068,7 @@ const elQueryBodyBuilder = async (context, user, options) => {
   const dateFilter = [];
   const searchAfter = after ? cursorToOffset(after) : undefined;
   let ordering = [];
-  const { adminBypassUserAccess = true } = options;
+  const { adminBypassUserAccess = false } = options;
   const markingRestrictions = await buildDataRestrictions(context, user, { adminBypassUserAccess });
   const accessMust = markingRestrictions.must;
   const accessMustNot = markingRestrictions.must_not;
