+++ opencti-platform/opencti-graphql/src/modules/workspace/workspace-domain.ts
@@ -20,7 +20,6 @@ import type { AuthContext, AuthUser } from '../../types/user';
 import type {
   EditContext,
   EditInput,
-  MemberAccess,
   MemberAccessInput,
   QueryWorkspacesArgs,
   StixRefRelationshipAddInput,
@@ -28,14 +27,11 @@ import type {
   WorkspaceAddInput,
   WorkspaceObjectsArgs
 } from '../../generated/graphql';
-import { findAllMembers } from '../../domain/user';
 import {
-  validateUserAccessOperation,
   MEMBER_ACCESS_RIGHT_ADMIN,
   isValidMemberAccessRight,
   getUserAccessRight
 } from '../../utils/access';
-import type { BasicStoreEntity } from '../../types/store';
 import { publishUserAction } from '../../listener/UserActionListener';
 
 export const findById = (context: AuthContext, user: AuthUser, workspaceId: string): BasicStoreEntityWorkspace => {
@@ -45,28 +41,6 @@ export const findById = (context: AuthContext, user: AuthUser, workspaceId: stri
 export const findAll = (context: AuthContext, user: AuthUser, args: QueryWorkspacesArgs) => {
   const listArgs = { ...args, adminBypassUserAccess: args.adminBypassUserAccess ?? false };
   return listEntitiesPaginated<BasicStoreEntityWorkspace>(context, user, [ENTITY_TYPE_WORKSPACE], listArgs);
-};
-
-export const getAuthorizedMembers = async (context: AuthContext, user: AuthUser, workspace: BasicStoreEntityWorkspace): Promise<MemberAccess[]> => {
-  let authorizedMembers: MemberAccess[] = [];
-  if (!workspace.authorized_members?.length) {
-    return authorizedMembers;
-  }
-  if (!validateUserAccessOperation(user, workspace, 'manage-access')) {
-    return authorizedMembers; // return empty if user doesn't have the right access_right
-  }
-  const membersIds = workspace.authorized_members.map((e) => e.id);
-  const args = {
-    connectionFormat: false,
-    first: 100,
-    filters: [{ key: 'internal_id', values: membersIds }],
-  };
-  const members = await findAllMembers(context, user, args);
-  authorizedMembers = workspace.authorized_members.map((am) => {
-    const member = members.find((m) => (m as BasicStoreEntity).id === am.id) as BasicStoreEntity;
-    return { id: am.id, name: member?.name ?? '', entity_type: member?.entity_type ?? '', access_right: am.access_right };
-  });
-  return authorizedMembers;
 };
 
 export const editAuthorizedMembers = async (context: AuthContext, user: AuthUser, workspaceId: string, input: MemberAccessInput[]) => {
