+++ opencti-platform/opencti-graphql/src/modules/workspace/workspace-resolver.ts
@@ -4,7 +4,6 @@ import {
   editAuthorizedMembers,
   findAll,
   findById,
-  getAuthorizedMembers,
   getCurrentUserAccessRight,
   getOwnerId,
   objects,
@@ -23,6 +22,7 @@ import { ENTITY_TYPE_WORKSPACE } from './workspace-types';
 import type { Resolvers } from '../../generated/graphql';
 import { batchLoader } from '../../database/middleware';
 import { batchCreator } from '../../domain/user';
+import { getAuthorizedMembers } from '../../domain/authorizedMembers';
 
 const creatorLoader = batchLoader(batchCreator);
