+++ opencti-platform/opencti-graphql/src/utils/access.ts
@@ -238,7 +238,7 @@ export const computeUserMemberAccessIds = (user: AuthUser) => {
 
 // user access methods
 export const getUserAccessRight = (user: AuthUser, element: any) => {
-  if (isBypassUser(user) || !element.authorized_members) { // no restricted user access on element
+  if (isUserHasCapability(user, SETTINGS_SET_ACCESSES) || !element.authorized_members) { // no restricted user access on element
     return MEMBER_ACCESS_RIGHT_ADMIN;
   }
   const accessMembers = [...element.authorized_members];
