+++ opencti-platform/opencti-graphql/tests/02-integration/02-resolvers/role-test.js
@@ -66,7 +66,7 @@ describe('Role resolver standard behavior', () => {
   });
   it('should list roles', async () => {
     const queryResult = await queryAsAdmin({ query: LIST_QUERY, variables: { first: 10 } });
-    expect(queryResult.data.roles.edges.length).toEqual(6);
+    expect(queryResult.data.roles.edges.length).toEqual(7);
   });
   it('should list capabilities', async () => {
     const LIST_CAPABILITIES_QUERY = gql`
